requirements.txt
openai>=1.79.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
langchain-openai>=0.3.17
tiktoken>=0.9.0
pydantic>=2.11.4 